requirements.txt
# Requirements for Darcy Chatbot Testing Project

# Web automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Testing framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0

# HTTP requests
requests>=2.31.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and utilities
coloredlogs>=15.0.1

# Optional: For more advanced web scraping if needed
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: For API testing
httpx>=0.24.0

# Web control interface
Flask>=3.0.0
flask-cors>=4.0.0
cryptography>=41.0.0

# Configuration
PyYAML>=6.0.0